Conflicts:
	src/Org.NerdBeers/Org.NerdBeers.Specs/Modules/SpecBootStrapper.cs
	src/Org.NerdBeers/Org.NerdBeers.Specs/Org.NerdBeers.Specs.csproj

+++ src/Org.NerdBeers/Org.NerdBeers.Web/Models/Comment.cs
@@ -7,9 +7,9 @@ namespace Org.NerdBeers.Web.Models
 {
     public class Comment
     {
-        public int Id { get; set; }
-        public int NerdId { get; set; }
-        public int EventId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64 NerdId { get; set; }
+        public Int64 EventId { get; set; }
         public string CommentText { get; set; }
         public DateTime Created {get;set;}
     }
